+++ lib/components/Message.jsx
@@ -1,0 +1,1 @@
+import React from 'react';
