+++ lib/components/MessageBox.jsx
@@ -1,0 +1,22 @@
+import React from 'react';
+
+
+class MessageBox extends React.Component {
+  constructor(){
+    super();
+  }
+
+  render() {
+
+  return (
+
+      <ul>
+        { this.props.messages.map(m => <li key={m.key}>{m.user.displayName}: {m.content}</li>) }
+
+      </ul>
+
+      );
+    }
+  }
+
+module.exports = MessageBox;
