+++ lib/components/Sort.jsx
@@ -1,0 +1,1 @@
+import React from 'react';
