+++ lib/index.js
@@ -1,10 +1,9 @@
 import React from 'react';
 import ReactDOM from 'react-dom';
-// import firebase from './firebase';
 import Application from './components/Application';
 
 require('./style.scss');
 
 ReactDOM.render(
   <Application />, document.getElementById('application')
-)
+);
